To split into pages
This is reducing the number of articles by 200 odd articles, because of
an encoding problem. Any suggestions on how to fix that?

Basically line 18 is where the problem is happening cause the line
cannot be decoded as utf-8. However that shouldnt result is losing the
entire file, but its hapenning. Might have to look into the previous
combination of files and see if I need to encode that properly